core: split up core
